+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - add latest getInviteUserDetails API, which can be called by specifying function parameter callLatestAPI: true
 
+### Changed
+- support trailing slash on Cidaas options: 'authority'
+
 ## V4.0.2
 
 ### Fixed